Rename the class.
