+++ osu.Game.Rulesets.Karaoke/Edit/Lyrics/Extends/ReferenceSettings.cs
@@ -1,13 +1,13 @@
-﻿// Copyright (c) andy840119 <andy840119@gmail.com>. Licensed under the GPL Licence.
+// Copyright (c) andy840119 <andy840119@gmail.com>. Licensed under the GPL Licence.
 // See the LICENCE file in the repository root for full licence text.
 
 using System.Collections.Generic;
 using osu.Framework.Graphics;
-using osu.Game.Rulesets.Karaoke.Edit.Lyrics.Extends.Texting;
+using osu.Game.Rulesets.Karaoke.Edit.Lyrics.Extends.Reference;
 
 namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics.Extends
 {
-    public class TextingExtend : EditExtend
+    public class ReferenceSettings : LyricEditorSettings
     {
         public override ExtendDirection Direction => ExtendDirection.Right;
 
@@ -15,8 +15,9 @@ namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics.Extends
 
         protected override IReadOnlyList<Drawable> CreateSections() => new Drawable[]
         {
-            new TextingEditModeSection(),
-            new TextingSwitchSpecialActionSection()
+            new ReferenceLyricAutoGenerateSection(),
+            new ReferenceLyricSection(),
+            new ReferenceLyricConfigSection()
         };
     }
 }
